// src/types/navProps.interface.ts
export interface PropsNav{
    backgroundColor?: string;
    padding?: string;
    margin?: string;
};

export interface PropsNavElement{
    width?: string;
    height?: string;
    moveTo?: string;
    left?: string;
    right?: string;
    flex?: string;
};

export interface PropsActiveLink{
    active: boolean;
    activeColor: string;
    passiveColor: string;
}

export interface PropsVisibile{
    visible: boolean;
}